feat: add testapp
