+++ .editorconfig
@@ -1,15 +1,15 @@
-# This file is for unifying the coding style of different editors and IDEs.
-# editorconfig.org
+# https://editorconfig.org
 
 root = true
 
 [*]
 charset = utf-8
-end_of_line = lf
-indent_size = 4
 indent_style = space
+indent_size = 2
+end_of_line = lf
 insert_final_newline = true
 trim_trailing_whitespace = true
 
-[*.json]
-indent_size = 2
+[*.md]
+insert_final_newline = false
+trim_trailing_whitespace = false
