--- .github/ISSUE_TEMPLATE.md
@@ -1,26 +1,0 @@
-### Expected Behaviour
-
-### Actual Behaviour
-
-### Reproduce Scenario (including but not limited to)
-
-#### Steps to Reproduce
-
-#### Platform and Version (eg. Android 5.0 or iOS 9.2.1)
-
-#### (Android) What device vendor (e.g. Samsung, HTC, Sony...)
-
-#### Cordova CLI version and cordova platform version
-
-    cordova --version                                    # e.g. 6.0.0
-    cordova platform version android                     # e.g. 4.1.1
-
-#### Plugin version
-
-    cordova plugin version | grep phonegap-plugin-speech-synthesis    # e.g. 1.5.3
-
-#### Sample Push Data Payload
-
-#### Sample Code that illustrates the problem
-
-#### Logs taken while reproducing problem
