--- .jshintrc
@@ -1,30 +1,0 @@
-{
-    "asi": false,
-    "boss": false,
-    "camelcase": true,
-    "curly": true,
-    "eqeqeq": true,
-    "eqnull": false,
-    "es5": false,
-    "evil": false,
-    "expr": false,
-    "forin": true,
-    "funcscope": false,
-    "jasmine": true,
-    "immed": true,
-    "indent": 4,
-    "latedef": true,
-    "loopfunc": false,
-    "maxerr": 7,
-    "newcap": true,
-    "node": true,
-    "nonew": true,
-    "plusplus": false,
-    "quotmark": "single",
-    "shadow": false,
-    "strict": false,
-    "supernew": false,
-    "trailing": true,
-    "undef": true,
-    "white": true
-}
