+++ .pre-commit-config.yaml
@@ -1,0 +1,22 @@
+repos:
+  - repo: https://github.com/pre-commit/pre-commit-hooks
+    rev: v4.5.0
+    hooks:
+      - id: trailing-whitespace
+      - id: end-of-file-fixer
+      - id: check-yaml
+  - repo: https://github.com/pre-commit/mirrors-prettier
+    rev: v3.0.3
+    hooks:
+      - id: prettier
+        types_or: [javascript]
+  - repo: https://github.com/pre-commit/mirrors-eslint
+    rev: v8.52.0
+    hooks:
+      - id: eslint
+        types: [file]
+        files: \.[jt]sx?$ # *.js, *.jsx, *.ts and *.tsx
+        args: [--cache, --max-warnings, "0"]
+        additional_dependencies:
+          - "eslint@8.52.0"
+          - "eslint-config-prettier@9.0.0"
