+++ package.json
@@ -1,10 +1,14 @@
 {
   "name": "@red-mobile/cordova-plugin-speech-synthesis",
-  "version": "0.2.0",
+  "version": "0.3.0",
   "description": "W3C Web Speech API - Speech synthesis plugin for Cordova",
   "homepage": "https://github.com/okhiroyuki/cordova-plugin-speech-synthesis#readme",
   "scripts": {
-    "test": "jasmine-node --color spec"
+    "test:clean": "rimraf testapp/node_modules testapp/plugins testapp/platforms testapp/temp-module-copy",
+    "test:install": "cd testapp && npm i",
+    "test:android": "cd testapp && cordova platform add android && cordova prepare android && cordova build",
+    "test:build": "run-s test:clean test:install test:android",
+    "lint": "npx eslint --cache --fix www && prettier --write www"
   },
   "repository": {
     "type": "git",
@@ -16,21 +20,20 @@
   "cordova": {
     "id": "@red-mobile/cordova-plugin-speech-synthesis",
     "platforms": [
-      "ios",
-      "android",
-      "browser"
+      "android"
     ]
   },
   "keywords": [
     "ecosystem:cordova",
-    "cordova-ios",
-    "cordova-android",
-    "cordova-browser"
+    "cordova-android"
   ],
-  "author": "Simon MacDonald",
+  "author": "Hiroyuki Okada",
   "license": "MIT",
-  "dependencies": {
-    "jasmine-node": "^1.14.5",
-    "pluginpub": "0.0.8"
+  "devDependencies": {
+    "eslint": "8.52.0",
+    "eslint-config-prettier": "9.0.0",
+    "npm-run-all2": "6.1.1",
+    "prettier": "3.0.3",
+    "rimraf": "5.0.5"
   }
 }
