+++ www/SpeechSynthesisEvent.js
@@ -1,7 +1,7 @@
-var SpeechSynthesisEvent = function() {
-    this.charIndex;
-    this.elapsedTime;
-    this.name;
+var SpeechSynthesisEvent = function () {
+  this.charIndex;
+  this.elapsedTime;
+  this.name;
 };
 
 module.exports = SpeechSynthesisEvent;
