+++ www/SpeechSynthesisVoiceList.js
@@ -1,11 +1,10 @@
-
-var SpeechSynthesisVoiceList = function(data) {
+var SpeechSynthesisVoiceList = function (data) {
   this._list = data;
   this.length = this._list.length;
 };
-    
-SpeechSynthesisVoiceList.prototype.item = function(item) {
-    return this._list[item];
+
+SpeechSynthesisVoiceList.prototype.item = function (item) {
+  return this._list[item];
 };
 
 module.exports = SpeechSynthesisVoiceList;
